refactor(key-value-operations): changed the naming convention
